services/mplus-svc/pyproject.toml
[tool.poetry]
name = "mplus-svc"
version = "0.1.0"
description = ""
authors = ["Victor Tantillo <v.tantillo3@gmail.com>"]
readme = "README.md"
packages = [{include = "mplus_svc"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
fastapi = {extras = ["all"], version = "^0.95.0"}
databases = {extras = ["asyncpg"], version = "^0.7.0"}
psycopg2 = "^2.9.5"
pyhumps = "^3.8.0"
ulid-py = "^1.1.0"
requests = "^2.28.2"
requests-toolbelt = "^0.10.1"
requests-oauthlib = "^1.3.1"
pytz = "^2022.7.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"

[tool.black]
line-length = 80

[tool.pyright]
reportMissingImports = true
useLibraryCodeForTypes = true
typeCheckingMode = "basic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
